Convert to using char for the primary key. Encapsulate all position fields of Page. Junit version to 4.11
git-svn-id: http://svn.mkgmap.org.uk/mkgmap/branches/unicode@3275 25d90789-57f7-4ee0-8453-03a3dfeeeb22

+++ src/uk/me/parabola/imgfmt/app/srt/SRTFile.java
@@ -101,7 +101,7 @@ public class SRTFile extends ImgFile {
 		int size = sort.getExpansionSize();
 		for (int j = 1; j <= size; j++) {
 			CodePosition b = sort.getExpansion(j);
-			writer.put(b.getPrimary());
+			writer.put((byte) b.getPrimary());
 			writer.put((byte) ((b.getTertiary() << 4) | (b.getSecondary() & 0xf)));
 		}
